@@ -31,8 +31,7 @@ class EagerSpanScanner extends SpanScanner {
 
   @override
   set state(LineScannerState state) {
-    if (state is! _EagerSpanScannerState ||
-        !identical((state as _EagerSpanScannerState)._scanner, this)) {
+    if (state is! _EagerSpanScannerState || !identical(state._scanner, this)) {
       throw ArgumentError('The given LineScannerState was not returned by '
           'this LineScanner.');
     }
@@ -69,7 +68,7 @@ class EagerSpanScanner extends SpanScanner {
     }
   }
 
-  EagerSpanScanner(String string, {sourceUrl, int position})
+  EagerSpanScanner(String string, {sourceUrl, int? position})
       : super(string, sourceUrl: sourceUrl, position: position);
 
   @override
@@ -99,13 +98,14 @@ class EagerSpanScanner extends SpanScanner {
   @override
   bool scan(Pattern pattern) {
     if (!super.scan(pattern)) return false;
+    final firstMatch = (lastMatch![0])!;
 
-    final newlines = _newlinesIn(lastMatch[0]);
+    final newlines = _newlinesIn(firstMatch);
     _line += newlines.length;
     if (newlines.isEmpty) {
-      _column += lastMatch[0].length;
+      _column += firstMatch.length;
     } else {
-      _column = lastMatch[0].length - newlines.last.end;
+      _column = firstMatch.length - newlines.last.end;
     }
 
     return true;
